rename 20240101_SMTC
